Try intp

+++ pandas/tests/io/formats/test_info.py
@@ -499,5 +499,5 @@ def test_memory_usage_empty_no_warning():
     df = DataFrame(index=["a", "b"])
     with tm.assert_produces_warning(None):
         result = df.memory_usage()
-    expected = Series(16, index=["Index"], dtype="int64")
+    expected = Series(16, index=["Index"])
     tm.assert_series_equal(result, expected)
